chore: structure testUtils better + move common mocks to separate file
