+++ tests/unit/botJoined.test.ts
@@ -1,3 +1,5 @@
+import "@/testUtils/mocks/mockEventBridge";
+
 import {
   EventBridgeClient,
   PutEventsCommand,
@@ -6,7 +8,7 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import type { Events } from "@/events";
 import { handler } from "@/functions/events/botJoined";
-import { mockEventBridgePayload } from "@/testUtils/mockEventBridgePayload";
+import { mockEventBridgePayload } from "@/testUtils/mocks/mockEventBridgePayload";
 import { sendMockSqsMessage } from "@/testUtils/sendMockSqsMessage";
 
 const constants = vi.hoisted(() => ({
@@ -14,18 +16,6 @@ const constants = vi.hoisted(() => ({
   coreChannelId: "C001",
   eventId: "E001",
 }));
-
-vi.mock("@aws-sdk/client-eventbridge", async () => {
-  const EventBridgeClient = vi.fn();
-  EventBridgeClient.prototype.send = vi.fn();
-
-  const PutEventsCommand = vi.fn();
-
-  return {
-    EventBridgeClient,
-    PutEventsCommand,
-  };
-});
 
 vi.mock("@/services/slack/getChannelMembers", () => ({
   getChannelMembers: vi.fn().mockResolvedValue(constants.channelMembers),
