+++ tests/utils/mocks/mockDb.ts
@@ -1,0 +1,5 @@
+import { vi } from "vitest";
+
+vi.mock("@/db/index", async () => ({
+  db: await import("@/testUtils/testDb").then(({ testDb }) => testDb),
+}));
