+++ tests/utils/mocks/mockLambdaPayload.ts
@@ -1,0 +1,42 @@
+export const mockLambdaEvent = {
+  body: "",
+  headers: {},
+  isBase64Encoded: false,
+  rawPath: "",
+  rawQueryString: "",
+  requestContext: {
+    accountId: "",
+    apiId: "",
+    domainName: "",
+    domainPrefix: "",
+    http: {
+      method: "",
+      path: "",
+      protocol: "",
+      sourceIp: "",
+      userAgent: "",
+    },
+    requestId: "",
+    routeKey: "",
+    stage: "",
+    time: "",
+    timeEpoch: 0,
+  },
+  routeKey: "",
+  version: "",
+};
+
+export const mockLambdaContext = {
+  awsRequestId: "",
+  callbackWaitsForEmptyEventLoop: false,
+  functionName: "",
+  functionVersion: "",
+  invokedFunctionArn: "",
+  logGroupName: "",
+  logStreamName: "",
+  memoryLimitInMB: "",
+  getRemainingTimeInMillis: () => 0,
+  done: () => {},
+  fail: () => {},
+  succeed: () => {},
+};
