+++ tests/utils/mocks/mockSlackApp.ts
@@ -1,0 +1,11 @@
+import { vi } from "vitest";
+
+vi.mock("@/services/slack/createSlackApp", () => ({
+  createSlackApp: vi.fn().mockReturnValue({
+    client: {
+      chat: {
+        postMessage: vi.fn(),
+      },
+    },
+  }),
+}));
